requirements.txt
fastapi==0.115.13
uvicorn[standard]==0.34.3
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
pydantic-settings==2.10.1